requirements.txt
blinker==1.8.2
certifi==2024.7.4
chardet==3.0.4
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
et-xmlfile==1.1.0
filelock==3.15.4
flask==3.0.3
Flask-JWT-Extended==4.6.0
fsspec==2024.6.1
googletrans==3.0.0
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==8.2.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pillow==10.4.0
PyJWT==2.8.0
pymongo==4.8.0
pytesseract==0.3.7
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2020.10.15
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tbb==2021.13.0
torch==2.3.1
torchvision==0.18.1
typing-extensions==4.12.2
tzdata==2024.1
waitress==3.0.0
werkzeug==3.0.3
zipp==3.19.2
